Add the runtime this stub is based on
For general sanity this has the runtime hex from the commit that the
stub was based on.

Practically speaking this means if you use the stub to build something
and then uflash updates you can continue to use the version of the API
your program was based on.

If you inspect the file, you may notice that it is not actually binary
and has a trailing newline on it. I don't know why this is, but it's
what uflash expects so this is what you get.